move s_main_game_globals and related functions to game module
