+++ xlive/Blam/Engine/game/aim_assist.cpp
@@ -2,7 +2,7 @@
 #include "aim_assist.h"
 
 #include "H2MOD.h"
-#include "Blam/Engine/game/game_globals.h"
+#include "Blam/Engine/game/game.h"
 #include "Util/Hooks/Hook.h"
 
 // sword-flying target clear patch
