+++ xlive/H2MOD/Modules/EventHandler/EventHandler.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "Blam/Engine/game/game_globals.h"
+#include "Blam/Engine/game/game_options.h"
 #include "Blam/Engine/Networking/logic/life_cycle_manager.h"
 #include "H2MOD/Modules/Shell/ServerConsole.h"
