+++ xlive/H2MOD/Modules/PlayerRepresentation/PlayerRepresentation.cpp
@@ -1,8 +1,9 @@
 #include "stdafx.h"
-
 #include "PlayerRepresentation.h"
+
 #include "Blam/Cache/TagGroups/biped_definition.hpp"
 #include "Blam/Cache/TagGroups/model_definition.hpp"
+#include "Blam/Engine/game/game.h"
 #include "Blam/Engine/game/game_engine.h"
 #include "Blam/Engine/game/game_engine_util.h"
 #include "Blam/Engine/game/players.h"
