+++ xlive/H2MOD/Modules/WeaponOffsets/WeaponOffsets.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 
-#include "Blam/Engine/game/game_globals.h"
+#include "Blam/Engine/game/game.h"
 #include "H2MOD/Modules/Input/PlayerControl.h"
 #include "H2MOD/GUI/ImGui_Integration/ImGui_Handler.h"
 #include "H2MOD/Tags/MetaLoader/tag_loader.h"
