+++ xlive/H2MOD/Tags/MetaLoader/tag_loader.cpp
@@ -2,7 +2,7 @@
 
 #include "tag_loader.h"
 
-#include "Blam/Engine/game/game_globals.h"
+#include "Blam/Engine/game/game_options.h"
 #include "Blam/Cache/TagGroups/model_definition.hpp"
 #include "Blam/Cache/TagGroups/scenery_definition.hpp"
 #include "H2MOD/Modules/OnScreenDebug/OnscreenDebug.h"
