+++ xlive/H2MOD/Variants/Infection/Infection.cpp
@@ -1,10 +1,9 @@
 #include "stdafx.h"
-
 #include "Infection.h"
 
 #include "Blam/Cache/TagGroups/item_collection_definition.hpp"
 #include "Blam/Cache/TagGroups/vehicle_collection_definition.hpp"
-#include "Blam/Engine/game/game_globals.h"
+#include "Blam/Engine/game/game.h"
 #include "Blam/Engine/game/game_time.h"
 #include "Blam/Engine/scenario/scenario.h"
 #include "Blam/Engine/Networking/logic/life_cycle_manager.h"
